requirements.txt
openai-whisper
python-dotenv
sounddevice
scipy
pydub

langchain
langchain_community
openai
langchain_openai

fastapi
uvicorn
pydantic
pyjwt
pymysql
python-jose